chore: organized admin codebase into components
