+++ acses-website/src/pages/admin/views/AdminEvents.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useState } from "react";
-import { useAdmin } from "./AdminContext";
-import { exportToCsv, formatDate } from "./adminUtils";
-import { fetchApi, unwrapList } from "../../utils/api";
+import { useAdmin } from "../context/AdminContext";
+import { exportToCsv, formatDate } from "../lib/adminUtils";
+import { fetchApi, unwrapList } from "../../../utils/api";
 
 const categories = ["seminar", "workshop", "social"];
 const statuses = ["upcoming", "ongoing", "past"];
