+++ acses-website/src/pages/admin/views/AdminUsers.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
-import { useAdmin } from "./AdminContext";
-import { exportToCsv } from "./adminUtils";
-import { fetchApi } from "../../utils/api";
+import { useAdmin } from "../context/AdminContext";
+import { exportToCsv } from "../lib/adminUtils";
+import { fetchApi } from "../../../utils/api";
 
 const roles = ["super admin", "editor"];
 const statuses = ["active", "disabled"];
